Modificação para permitir alterar e excluir tarefas separadas por usuário

+++ gerenciador_tarefas/app/views/tarefa_views.py
@@ -1,5 +1,5 @@
 from django.contrib.auth.decorators import login_required
-from django.shortcuts import render, redirect
+from django.shortcuts import render, redirect, HttpResponse
 from ..forms import TarefaForm
 from ..entidades.tarefa import Tarefa
 from ..services import tarefa_service
@@ -31,12 +31,14 @@ def cadastrar_tarefa(request):
             return redirect('listar_tarefas')
     else:
         form_tarefa = TarefaForm()
-    return render(request, 'tarefas/form_tarefa.html', {'form_tarefa':form_tarefa})
+    return render(request, 'tarefas/form_tarefa.html', {'form_tarefa': form_tarefa})
 
 
 @login_required
 def editar_tarefa(request, id):
     tarefa_bd = tarefa_service.lista_tarefa_id(id)
+    if tarefa_bd.usuario != request.user:
+        return HttpResponse('Não permitido!')
     form_tarefa = TarefaForm(request.POST or None, instance=tarefa_bd)
     if form_tarefa.is_valid():
         titulo = form_tarefa.cleaned_data['titulo']
@@ -52,14 +54,15 @@ def editar_tarefa(request, id):
                         )
         tarefa_service.editar_tarefa(tarefa_bd, tarefa_nova)
         return redirect('listar_tarefas')
-    return render(request, 'tarefas/form_tarefa.html', {'form_tarefa':form_tarefa})
+    return render(request, 'tarefas/form_tarefa.html', {'form_tarefa': form_tarefa})
 
 
 @login_required
 def remover_tarefa(request, id):
     tarefa_bd = tarefa_service.lista_tarefa_id(id)
+    if tarefa_bd.usuario != request.user:
+        return HttpResponse('Não permitido!')
     if request.method == "POST":
         tarefa_service.remover_tarefa(tarefa_bd)
         return redirect('listar_tarefas')
-    return render(request, 'tarefas/confirma_exclusao.html', {'tarefa':tarefa_bd})
-
+    return render(request, 'tarefas/confirma_exclusao.html', {'tarefa': tarefa_bd})
